- Updated cores to current. - Mightiest Mark Dates are now checked through ALM. - Minor cleaning. - Refactor embed images. - Added Italian to SR - Readded ISO Language codes - Refactor Medal images, preparing to swap. - Fuckling with footer texts because nitpicking.

+++ SysBot.Pokemon.Discord/Commands/Bots/SpecialRequestModule.cs
@@ -145,7 +145,7 @@ namespace SysBot.Pokemon.Discord
                     return;
                 }
 
-                var code = Info.GetRandomTradeCode((int)userID);
+                var code = Info.GetRandomTradeCode(userID);
                 var lgcode = Info.GetRandomLGTradeCode();
                 var sig = Context.User.GetFavor();
